+++ objectified-ui/public/WHATS_NEW.md
@@ -38,6 +38,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Fixed JSON Schema refs, was #/components/schemas, but this needs to be #/$defs according to the spec.
   - Fixed Canvas Layout option so that Load is only available if a previous layout has been saved.
   - Fixed issue with system theming not applying properly on initial load.
+  - Fixed edges in canvas so they don't overlap over the top of classes, but go underneath.
   - Compacted the look and feel of the Class nodes to reduce visual clutter and improve readability.
   - Class Template Library now shows the schema preview.
   - Properties and Class Template Form now includes a "close" button